Hopefully fix overflow issue with BTsolver number of branhces checked and fix problem with max degree heuristic

+++ src/BTSolver.java
@@ -2,7 +2,7 @@
 public class BTSolver {
 
     BinaryCSP problem;
-    int nodesExplored;
+    long nodesExplored;
 
     public BTSolver(BinaryCSP problem) {
